Add a category as screen
